doc(drone_decorator_test.cc, robot_decorator_test.cc, etc..):#11 Added Delivery object decorator tests and fixed test compile errors resulting from changes to other code.
Created tests for both robot and drone decorators as well as fixed all of the
compile errors for tests as a result of the changes to drone and robot and the
additions of basic drone and basic robot. These changes simply changed the types
slightly and the names so that it would be basic_robots and basic_drones rather
then the error caused by trying to create the now abstract robot and drone types.

Changes:
-Renamed: drone_test.cc to basic_drone_test.cc

-Renamed: package_test.cc to basic_package_test.cc

-Renamed: robot_test.cc to basic_robot_test.cc

New:

robot_decorator_test.cc
-Added test to see if the decorator actually returns the data of the original robot

drone_decorator_test.cc
-Added test to see if the decorator actually returns the data of the original drone

Bugfixes:

fixed all bugs resulting in compile errors in make test.