doc(customer_color_decorator_test.cc): #13 customer color decorator test was added.
Tests for customer color decorator were added as well as fixes for customer factory test
and delivery object test in order to allow them to work with the changes to customer. To go
along with that customer test was renamed to basic customer test and now uses basic customer
as customer is now abstract.